+++ src/views/components/molecules/Layout/AdminPageLayout.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+// import Button from '../../atoms/Button';
+import Navbar from '../Navbar';
+
+type Props = {
+  children: React.ReactNode;
+};
+
+const AdminPageLayout: React.FC<Props> = ({ children }) => {
+  return (
+    <>
+      <section className="w-full py-28 md:py-0 FlexCenter flex-col bg-gradient-to-b from-Pink-lighter to-Pink-default md:h-screen relative text-Pink-default">
+        <Navbar
+          bgColor="Pink-default"
+          link={
+            <li className="mr-4 Hover">
+              <Link to="/admin/event">Events</Link>
+            </li>
+          }
+          redirectPath="/admin/login"
+        />
+        {children}
+        <small className="absolute bottom-5 text-center w-full">
+          &copy; Sho, Kyosuke, Fumina, Ayaka 2022 / All Rights Reserved
+        </small>
+      </section>
+    </>
+  );
+};
+
+export default AdminPageLayout;
